Setup system for wait()
thread and process id are now different
for each process we save a state, like which children exists
and which may be wait()ed on. also save process state, like
normal or zombie (not waited on).

implementation is good enough for a single program (like echo)
but is not complete