Use frame table from previous commit.
The user pool from palloc.c is not used anymore.
Assert this for all calls.
Instead use frame_get_free() to get a page from the
user pool.
Change process.c to use the new allocator.

TODO:
probably best to completely hide the new frame allocator and
use it under the hood if PAL_USER is given to any call of
palloc.c