Differentiate handling of segments and mmap.
While both read from a file with offset, only mmap-ed
files must be written back. A segment, even if the page is
writable, will never be written back to the file, it may
only be swapped. This is important for the destruction of
the sup. page table as only mmap-ed files need special handling.